fix: fix database syntax script error
