+++ src/Data/Interfaces/IDbConnectionFactory.cs
@@ -1,6 +1,6 @@
 using System.Data;
 
-namespace ExpenseTrackerGroup3.Data;
+namespace ExpenseTrackerGroup3.Data.Interfaces;
 
 public interface IDbConnectionFactory 
 {
