+++ src/Data/Interfaces/IDbInitializer.cs
@@ -1,0 +1,6 @@
+namespace ExpenseTrackerGroup3.Data.Interfaces;
+
+public interface IDbInitializer 
+{
+    public void InitializeDatabase();
+}
