+++ src/ExpenseTrackerGroup3.csproj
@@ -11,4 +11,7 @@
     <PackageReference Include="Npgsql" Version="8.0.4" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0" />
   </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="Data/Scripts/*.sql" />
+  </ItemGroup>
 </Project>
